Conflicts:
	ChangeLog
	ailurus/support/clientlib.py

+++ ailurus/ubuntu/apps.py
@@ -1,10 +1,9 @@
-#!/usr/bin/env python
-#-*- coding: utf-8 -*-
+#coding: utf8
 #
-# Ailurus - make Linux easier to use
+# Ailurus - a simple application installer and GNOME tweaker
 #
+# Copyright (C) 2009-2010, Ailurus developers and Ailurus contributors
 # Copyright (C) 2007-2010, Trusted Digital Technology Laboratory, Shanghai Jiao Tong University, China.
-# Copyright (C) 2009-2010, Ailurus Developers Team
 #
 # Ailurus is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -93,7 +92,7 @@ class GNOMEArtNextGen(I):
         import os
         path = os.path.expanduser('~/.gnome2/gnome-art-ng/')
         if not os.path.exists(path): run('mkdir '+path)
-        with Chdir(path) as o:
+        with Chdir(path):
             run('tar xf '+thumb)
     def installed(self):
         return APT.installed('gnomeartng')
@@ -109,37 +108,40 @@ class DisableGetty(I):
     def visible(self):
         return os.path.exists('/etc/event.d/tty1')
     def installed(self):
-        with Chdir('/etc/event.d/') as o:
+        with Chdir('/etc/event.d/'):
             for i in range(2,7):
                 file_name = 'tty%s' % i
-                with open(file_name) as f:
-                    for line in f:
-                        if line.startswith('exec'): return False
+                if os.path.exists(file_name):
+                    with open(file_name) as f:
+                        for line in f:
+                            if line.startswith('exec'): return False
         return True
     def install(self):
-        with Chdir('/etc/event.d/') as o:
+        with Chdir('/etc/event.d/'):
             for i in range(2,7):
                 file_name = 'tty%s'%i
-                with TempOwn(file_name) as o:
-                    with open(file_name) as f:
-                        contents = f.readlines()
-                    for j, line in enumerate(contents):
-                        if line.startswith('exec'):
-                            contents[j]='#' + line
-                    with open(file_name, 'w') as f:
-                        f.writelines(contents)
+                if os.path.exists(file_name):
+                    with TempOwn(file_name):
+                        with open(file_name) as f:
+                            contents = f.readlines()
+                        for j, line in enumerate(contents):
+                            if line.startswith('exec'):
+                                contents[j]='#' + line
+                        with open(file_name, 'w') as f:
+                            f.writelines(contents)
     def remove(self):
-        with Chdir('/etc/event.d/') as o:
+        with Chdir('/etc/event.d/'):
             for i in range(2,7):
                 file_name = 'tty%s'%i
-                with TempOwn(file_name) as o:
-                    with open(file_name) as f:
-                        contents = f.readlines()
-                    for j, line in enumerate(contents):
-                        if line.startswith('#exec'):
-                            contents[j]='exec /sbin/getty 38400 tty%s\n' % i
-                    with open(file_name, 'w') as f:
-                        f.writelines(contents)
+                if os.path.exists(file_name):
+                    with TempOwn(file_name):
+                        with open(file_name) as f:
+                            contents = f.readlines()
+                        for j, line in enumerate(contents):
+                            if line.startswith('#exec'):
+                                contents[j]='exec /sbin/getty 38400 tty%s\n' % i
+                        with open(file_name, 'w') as f:
+                            f.writelines(contents)
 
 class DisableGettyKarmic(I):
     __doc__ = _('Deactivate Getty ( Ctrl+Alt+F2 ... F6 ), Ctrl+Alt+F1 is still activated')
@@ -147,18 +149,19 @@ class DisableGettyKarmic(I):
     def visible(self):
         return os.path.exists('/etc/init/tty1.conf')
     def installed(self):
-        with Chdir('/etc/init/') as o:
+        with Chdir('/etc/init/'):
             for i in range(2,7):
                 file_name = 'tty%s.conf' % i
-                with open(file_name) as f:
-                    for line in f:
-                        if line.startswith('exec'): return False
+                if os.path.exists(file_name):
+                    with open(file_name) as f:
+                        for line in f:
+                            if line.startswith('exec'): return False
         return True
     def install(self):
-        with Chdir('/etc/init/') as o:
+        with Chdir('/etc/init/'):
             for i in range(2,7):
                 filename = 'tty%s.conf'%i
-                with TempOwn(filename) as o:
+                with TempOwn(filename):
                     with open(filename) as f:
                         contents = f.readlines()
                     for j, line in enumerate(contents):
@@ -167,10 +170,10 @@ class DisableGettyKarmic(I):
                     with open(filename, 'w') as f:
                         f.writelines(contents)
     def remove(self):
-        with Chdir('/etc/init/') as o:
+        with Chdir('/etc/init/'):
             for i in range(2,7):
                 filename = 'tty%s.conf'%i
-                with TempOwn(filename) as o:
+                with TempOwn(filename):
                     with open(filename) as f:
                         contents = f.readlines()
                     for j, line in enumerate(contents):
